+++ src/dipdup/config.py
@@ -970,10 +970,12 @@ default_hooks = {
 @dataclass
 class AdvancedConfig:
     reindex: Dict[ReindexingReasonC, ReindexingAction] = field(default_factory=dict)
+    # TODO: Drop in major version
     oneshot: bool = False
     postpone_jobs: bool = False
     early_realtime: bool = False
     merge_subscriptions: bool = False
+    metadata_interface: bool = False
 
 
 @dataclass