+++ src/dipdup/context.py
@@ -46,9 +46,11 @@ from dipdup.exceptions import IndexAlreadyExistsError
 from dipdup.exceptions import InitializationRequiredError
 from dipdup.exceptions import ReindexingRequiredError
 from dipdup.models import Contract
+from dipdup.models import ContractMetadata
 from dipdup.models import Index
 from dipdup.models import ReindexingReason
 from dipdup.models import Schema
+from dipdup.models import TokenMetadata
 from dipdup.prometheus import Metrics
 from dipdup.utils import FormattedLogger
 from dipdup.utils.database import execute_sql_scripts
@@ -56,6 +58,31 @@ from dipdup.utils.database import in_global_transaction
 from dipdup.utils.database import wipe_schema
 
 pending_indexes = deque()  # type: ignore
+
+
+class MetadataCursor:
+    _contract = 0
+    _token = 0
+
+    def __new__(cls):
+        raise NotImplementedError
+
+    @classmethod
+    async def initialize(cls) -> None:
+        if last_contract := await ContractMetadata.filter().order_by('-update_id').get_or_none():
+            cls._contract = last_contract.update_id
+        if last_token := await TokenMetadata.filter().order_by('-update_id').get_or_none():
+            cls._token = last_token.update_id
+
+    @classmethod
+    def contract(cls) -> int:
+        cls._contract += 1
+        return cls._contract
+
+    @classmethod
+    def token(cls) -> int:
+        cls._token += 1
+        return cls._token
 
 
 # TODO: Dataclasses are cool, everyone loves them. Resolve issue with pydantic serialization.
@@ -206,6 +233,25 @@ class DipDupContext:
 
         # NOTE: IndexDispatcher will handle further initialization when it's time
         pending_indexes.append(index)
+
+    async def update_contract_metadata(self, address: str, metadata: Dict[str, Any]) -> None:
+        if not self.config.advanced.metadata_interface:
+            return
+        update_id = MetadataCursor.contract()
+        await ContractMetadata.update_or_create(
+            contract=address,
+            defaults={'metadata': metadata, 'update_id': update_id},
+        )
+
+    async def update_token_metadata(self, address: str, token_id: int, metadata: Dict[str, Any]) -> None:
+        if not self.config.advanced.metadata_interface:
+            return
+        update_id = MetadataCursor.token()
+        await TokenMetadata.update_or_create(
+            contract=address,
+            token_id=token_id,
+            defaults={'metadata': metadata, 'update_id': update_id},
+        )
 
 
 class HookContext(DipDupContext):