+++ src/dipdup/http.py
@@ -12,6 +12,7 @@ from typing import Tuple
 from typing import cast
 
 import aiohttp
+import orjson  # type: ignore
 from aiolimiter import AsyncLimiter
 from fcache.cache import FileCache  # type: ignore
 
@@ -75,6 +76,7 @@ class _HTTPGateway:
     async def __aenter__(self) -> None:
         """Create underlying aiohttp session"""
         self.__session = aiohttp.ClientSession(
+            json_serialize=lambda *a, **kw: orjson.dumps(*a, **kw).decode(),
             connector=aiohttp.TCPConnector(limit=self._config.connection_limit or 100),
             timeout=aiohttp.ClientTimeout(connect=self._config.connection_timeout or 60),
         )